Fix remaining voice post duration type mismatch

+++ app/src/main/java/com/fynx/app/ui/FynxVoicePostRecorder.kt
@@ -102,7 +102,7 @@ fun FynxVoicePostRecorder(onRecorded: (Uri) -> Unit, onDismiss: () -> Unit) {
         hasRecording = outputFile?.let { it.exists() && it.length() > 0L } == true
         elapsedMs = elapsedMs.coerceAtMost(MAX_VOICE_POST_DURATION_MS)
         previewPlayer?.let {
-            previewDurationMs = it.duration.coerceAtLeast(0).toLong().toLong()
+            previewDurationMs = it.duration.coerceAtLeast(0)
             previewPositionMs = 0L
         }
     }
@@ -147,7 +147,7 @@ fun FynxVoicePostRecorder(onRecorded: (Uri) -> Unit, onDismiss: () -> Unit) {
             playing = false
             previewPositionMs = 0L
         }
-        previewPlayer?.let { previewDurationMs = it.duration.coerceAtLeast(0) }
+        previewPlayer?.let { previewDurationMs = it.duration.coerceAtLeast(0).toLong() }
         onDispose { previewPlayer?.runCatching { release() } }
     }
